ext-diet WP1+WP2: EnvoyExt layout geometry + Embot viz move to module DATs
EnvoyExt 9,221 -> 8,234 lines. envoy_layout.py (158) holds lint/dock-hug/
auto-position; envoy_viz.py (1,035) holds the entire Embot + camera-follow
subsystem (29 methods + _VIZ_* constants) behind delegating stubs, state
on the ext. Verified: 3-reviewer adversarial panel (AST-diffed bodies,
integration + hygiene lenses), live Embot smoke, 76 tests green, tool
schema byte-identical. Panel fixes: never-raises guard restored on the
_noteVizActivity chokepoint stub, dead colorsys/random imports dropped,
annotations no longer repel the auto-position scan. Placeholder modules
for WP3-5 wired (envoy_ops/read/setup). Plan: dev/embody/plan-ext-diet.md
