+++ dev/embody.tdn
@@ -1,10 +1,10 @@
 format: tdn
 version: '2.0'
-build: 499
-generator: Embody/6.0.100
+build: 503
+generator: Embody/6.0.104
 td_build: 099.2025.32820
-source_file: Embody-6.101.toe
-exported_at: '2026-07-04T20:26:33Z'
+source_file: Embody-6.104.toe
+exported_at: '2026-07-06T04:19:23Z'
 network_path: /embody
 options:
   include_dat_content: false
@@ -109,6 +109,7 @@ operators:
     - name: Refresh
       style: Pulse
       help: Refreshes the externalizations table by re-scanning tagged operators without triggering saves or continuity checks.
+      value: true
     - name: Performmode
       style: Toggle
       label: Perform Mode
@@ -304,7 +305,7 @@ operators:
       label: Auto-Save Status
       readOnly: true
       help: 'Read-only auto-save state: Idle / Saving / Bypassed (Perform Mode) / Disabled.'
-      value: Saved 17:30:27 UTC
+      value: Saved 04:19:18 UTC
     - name: Tdnstriponsave
       style: Toggle
       label: Strip on Save
@@ -673,7 +674,7 @@ operators:
       startSection: true
       readOnly: true
       help: Embody version string (read-only).
-      value: 6.0.100
+      value: 6.0.104
     - name: Touchbuild
       style: Str
       label: Touch Build
@@ -690,7 +691,7 @@ operators:
       label: Build Number
       readOnly: true
       help: Embody build number (read-only). Incremented with each release.
-      value: 716
+      value: 721
     - name: Github
       style: Pulse
       label: GitHub
@@ -701,7 +702,7 @@ operators:
       label: Build Date
       readOnly: true
       help: Date of this Embody build (read-only).
-      value: 2026-07-04 20:26:16 UTC
+      value: 2026-07-06 04:19:23 UTC
     - name: Help
       style: Pulse
       help: Opens the Embody help documentation.
@@ -720,15 +721,13 @@ operators:
     _start_in_progress: true
     git_status:
       /embody/externalizations: ' M'
-      /embody/Embody/templates/text_settings_local: ' M'
       /embody/Embody/EnvoyExt: ' M'
-      /embody/Embody/EmbodyExt: ' M'
       /embody/Embody: ' M'
       /embody: ' M'
-      /embody/Embody/wizard/logic: ??
-      /embody/Embody/wizard/clicks: ??
-      /embody/Embody/wizard: ??
-    _watchdog_gen: 743
+      /embody/Embody/envoy_layout: ??
+      /embody/Embody/envoy_viz: ??
+      /embody/Embody/envoy_ops: ??
+    _watchdog_gen: 757
     _suite_proof_progress:
       todo:
       - test_ancestor_rename
@@ -870,8 +869,8 @@ operators:
           tot: 33
           fail: 1
       pending_after: absent
-    _clip_watch_gen: 508
-    _suppress_dialogs: true
+    _clip_watch_gen: 522
+    _init_complete: true
   tdn_ref: embody/Embody.tdn
 - name: base_tdn
   type: baseCOMP
@@ -2797,7 +2796,7 @@ operators:
   parameters:
     file: embody/externalizations.tsv
     syncfile: true
-    rows: 263
+    rows: 268
     cols: 11
   flags:
   - viewer
