+++ dev/embody/Embody/envoy_ops.py
@@ -1,0 +1,4 @@
+"""Envoy mutating tool handlers (module DAT).
+
+Populated by the ext-diet refactor (WP3). Called by EnvoyExt only.
+"""
