+++ dev/embody/Embody/envoy_read.py
@@ -1,0 +1,4 @@
+"""Envoy read/introspection tool handlers (module DAT).
+
+Populated by the ext-diet refactor (WP4). Called by EnvoyExt only.
+"""
