+++ dev/embody/Embody/envoy_setup.py
@@ -1,0 +1,4 @@
+"""Envoy environment/config setup (module DAT).
+
+Populated by the ext-diet refactor (WP5). Called by EnvoyExt only.
+"""
