+++ dev/embody/externalizations.tsv
@@ -101,7 +101,7 @@ path	type	strategy	rel_file_path	timestamp	dirty	build	touch_build	node_x	node_y
 /embody/unit_tests/test_sandbox/sandbox_test_custom_parameters2	base	tdn	embody/unit_tests/test_sandbox/sandbox_test_custom_parameters2.tdn	2026-06-20 20:36:03 UTC		1	099.2025.32820	0	0	0.3000,0.5000,0.9000
 /embody/unit_tests/test_sandbox/sandbox_test_custom_parameters1	base	tdn	embody/unit_tests/test_sandbox/sandbox_test_custom_parameters1.tdn	2026-06-20 20:36:03 UTC		1	099.2025.32820	0	0	0.3000,0.5000,0.9000
 /embody/unit_tests/test_sandbox/sandbox_test_custom_parameters	base	tdn	embody/unit_tests/test_sandbox/sandbox_test_custom_parameters.tdn	2026-06-20 20:36:03 UTC		1	099.2025.32820	0	0	0.3000,0.5000,0.9000
-/embody/unit_tests/test_sandbox	base	tdn	embody/unit_tests/test_sandbox.tdn	2026-07-04 11:49:28 UTC		606	099.2025.32820	400	0	0.3000,0.5000,0.9000
+/embody/unit_tests/test_sandbox	base	tdn	embody/unit_tests/test_sandbox.tdn	2026-07-06 04:44:35 UTC		606	099.2025.32820	400	0	0.3000,0.5000,0.9000
 /embody/unit_tests/test_rename_move_lifecycle	text	py	embody/unit_tests/test_rename_move_lifecycle.py	2026-06-20 20:36:03 UTC				200	-600	0.9000,0.3000,0.4000
 /embody/unit_tests/test_path_utils	text	py	embody/unit_tests/test_path_utils.py	2026-06-20 20:36:03 UTC				200	-800	0.9000,0.3000,0.4000
 /embody/unit_tests/test_param_tracker	text	py	embody/unit_tests/test_param_tracker.py	2026-06-20 20:36:03 UTC				400	-800	0.9000,0.3000,0.4000
@@ -238,8 +238,8 @@ path	type	strategy	rel_file_path	timestamp	dirty	build	touch_build	node_x	node_y
 /embody/Embody/Collection/CollectionExt	text	py	embody/Embody/Collection/CollectionExt.py	2026-06-20 20:36:05 UTC				200	0	0.9000,0.3000,0.4000
 /embody/Embody/Collection	base	tdn	embody/Embody/Collection.tdn	2026-06-20 20:36:05 UTC		1	099.2025.32820	200	-550	0.3000,0.5000,0.9000
 /embody/Embody/CatalogManagerExt	text	py	embody/Embody/CatalogManagerExt.py	2026-06-20 20:36:05 UTC				-300	300	0.9000,0.3000,0.4000
-/embody/Embody	container	tdn	embody/Embody.tdn	2026-07-04 20:26:19 UTC		716	099.2025.32820	0	0	0.3000,0.5000,0.9000
-/embody	container	tdn	embody.tdn	2026-07-04 20:26:37 UTC		499	099.2025.32820	100	50	0.3000,0.5000,0.9000
+/embody/Embody	container	tdn	embody/Embody.tdn	2026-07-06 04:45:05 UTC		721	099.2025.32820	0	0	0.3000,0.5000,0.9000
+/embody	container	tdn	embody.tdn	2026-07-06 04:19:23 UTC		503	099.2025.32820	100	50	0.3000,0.5000,0.9000
 /embody/unit_tests/test_tdn_diff_engine	text	py	embody/unit_tests/test_tdn_diff_engine.py	2026-06-22 21:36:58 UTC				200	-3200	0.9000,0.3000,0.4000
 /embody/unit_tests/test_tdn_diff	text	py	embody/unit_tests/test_tdn_diff.py	2026-06-22 21:37:00 UTC				0	-3200	0.9000,0.3000,0.4000
 /embody/unit_tests/test_git_status	text	py	embody/unit_tests/test_git_status.py	2026-06-22 23:29:58 UTC				400	-3200	0.9000,0.3000,0.4000
@@ -261,3 +261,8 @@ path	type	strategy	rel_file_path	timestamp	dirty	build	touch_build	node_x	node_y
 /embody/Embody/wizard/logic	text	py	embody/Embody/wizard/logic.py	2026-07-04 17:29:32 UTC				240	0	0.9000,0.3000,0.4000
 /embody/Embody/wizard/clicks	panelexec	py	embody/Embody/wizard/clicks.py	2026-07-04 17:29:54 UTC				240	-150	0.9000,0.3000,0.4000
 /embody/Embody/wizard	container	tdn	embody/Embody/wizard.tdn	2026-07-04 20:26:13 UTC		2	099.2025.32820	1200	225	0.3000,0.5000,0.9000
+/embody/Embody/envoy_layout	text	py	embody/Embody/envoy_layout.py	2026-07-06 04:16:05 UTC				-500	500	0.9000,0.3000,0.4000
+/embody/Embody/envoy_viz	text	py	embody/Embody/envoy_viz.py	2026-07-06 04:19:16 UTC				-500	600	0.9000,0.3000,0.4000
+/embody/Embody/envoy_ops	text	py	embody/Embody/envoy_ops.py	2026-07-06 04:19:17 UTC				-500	400	0.9000,0.3000,0.4000
+/embody/Embody/envoy_read	text	py	embody/Embody/envoy_read.py	2026-07-06 04:19:22 UTC				-500	800	0.9000,0.3000,0.4000
+/embody/Embody/envoy_setup	text	py	embody/Embody/envoy_setup.py	2026-07-06 04:19:23 UTC				-500	700	0.9000,0.3000,0.4000
